+++ Backlog/CompletionDialog.xaml.cs
@@ -26,10 +26,10 @@ namespace Backlog
 
         private void TimeSpentTextBox_PreviewTextInput(object sender, TextCompositionEventArgs e)
         {
-            int a;
+            double a;
             try
             {
-                a = int.Parse(e.Text);
+                a = double.Parse(e.Text);
             }
             catch (Exception)
             {
